+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Replace `ChannelMetaData._typeGraph` boolean flag with `ChannelMetaData._instanceRootEntityTypeId` to make it easier to drive business logic and to write tests for subscription logic.
 * Extract generic `ReplicantSessionManagerImpl.processDeleteMessages()` from downstream libraries to make it easier to refactor logic in subscription managers.
 * Expose instance channels by index in `SystemMetaData`. The ultimate goal is to create a fast mechanism for accessing metadata in the subscription manager.
+* Implement previously abstract `ReplicantSessionManagerImpl.shouldFollowLink()` and generate an exception if called without being overriden.
 
 ### [v6.66](https://github.com/replicant4j/replicant/tree/v6.66) (2020-02-05) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.65...v6.66)